requirements.txt
numpy==1.22.2
pandas==1.4.1
python-dateutil==2.8.2
pytz==2021.3
scipy==1.8.0
six==1.16.0
tqdm==4.62.3
typing_extensions==4.1.1
scikit-learn==1.4.0
easydict==1.11
hydra-core==1.3.2
wandb==0.16.3
